+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Also, the following bug due to updating dependencies has been fixed:
 
 - fixed projection example in json schema generator
 - `serializerCompiler` has been added to use explicitly [`fastifiy-fast-json`](https://github.com/fastify/fast-json-stringify), along with `AdditionalCaster`  
+- [#326](https://github.com/mia-platform/crud-service/issues/326): `/schemas` and `/-/schemas` endpoints now return also `required` property
 
 ## 7.0.2 - 2024-05-06